vision-service/requirements.txt
# ETMS Vision Service - Dependencies
# Core ML
ultralytics>=8.3.0
torch>=2.0.0
torchvision>=0.15.0

# Computer Vision
opencv-python>=4.8.0
numpy>=1.24.0

# MQTT
paho-mqtt>=2.0.0

# Configuration
pyyaml>=6.0

# Geometry (for zone detection)
shapely>=2.0.0

# Utilities
pillow>=10.0.0
